+++ OpenGLLearning/OpenGLLearning/src/tests/TestTexture2D.cpp
@@ -53,7 +53,7 @@ namespace test
 		GLCall(glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA));
 
 		m_VAO = std::make_unique<VertexArray>();
-		m_VertexBuffer = std::make_unique<VertexBuffer>(positions, sizeof(positions));
+		m_VertexBuffer = std::make_unique<VertexBuffer>(positions, (unsigned int)sizeof(positions));
 		VertexBufferLayout layout;
 		layout.Push<float>(2);
 		layout.Push<float>(2);
